requirements.txt
numpy
pandas
yfinance
torch
scikit-learn
matplotlib
pyyaml
pyarrow